version v.0.0.3
Changed how it runs the minecraft server, instead of relying on the .bat file, it will now run the jar directly. Also added an about page to add some small insight along with a new logo for Osolo. Also removed a shit ton of useless pictures, decreasing the size of Osolo (hopefully).